+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
+## v.1.0.1
+    - add new consts for responses
+    - update parseResponse() method
+    - updae tests
 ## v.1.0.0
     - initial release